+++ src/bundler.ts
@@ -3,8 +3,8 @@ import { resolve } from 'pathe'
 import { extendWebpackConfig, extendViteConfig, addWebpackPlugin, addVitePlugin } from '@nuxt/kit'
 import VueI18nWebpackPlugin from '@intlify/unplugin-vue-i18n/webpack'
 import VueI18nVitePlugin from '@intlify/unplugin-vue-i18n/vite'
-import { TransformMacroPlugin, TransformMacroPluginOptions } from './macros'
-import { ResourceProxyPlugin } from './proxy'
+import { TransformMacroPlugin, TransformMacroPluginOptions } from './transform/macros'
+import { ResourceProxyPlugin } from './transform/proxy'
 
 import type { Nuxt } from '@nuxt/schema'
 import type { NuxtI18nOptions } from './types'
